Making sure that the project compiles on Windows when HAVE_MDNS_SUPPORT is present. Moving the 2 additional LRU functions from mdns.c to lruhash.c Defining the 2 additional functions in lruhash.h
